requirements.txt
bs4==0.0.2
chromadb==0.5.12
langgraph==0.2.35
langgraph-sdk==0.1.32
langgraph-checkpoint-sqlite==2.0.0
langchain-community==0.3.2
langchain-core==0.3.10
langchain-openai==0.2.2
langchain-postgres==0.0.12
langsmith==0.1.135
notebook
openai==1.51.2
psycopg-binary==3.2.3
python-dotenv==1.0.1
requests==2.32.3